Split pattern matching patches by class
