+++ lib/ruby-next/core/struct/deconstruct.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RubyNext::Core.patch Struct, method: :deconstruct, version: "2.7" do
+  <<~'RUBY'
+    alias deconstruct to_a
+  RUBY
+end
